requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.0
grpcio-tools==1.59.0
zmq==0.0.0
pyzmq==25.1.1
neo4j==5.14.1
networkx==3.2.1
torch==2.1.0
stable-baselines3==2.1.0
gymnasium==0.29.1
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0
aiohttp==3.9.0
celery==5.3.4
redis==5.0.1
docker==6.1.3
kubernetes==28.1.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.18.1
google-cloud-monitoring==2.16.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-nmap==0.7.1
shodan==1.30.1
censys==2.2.7
dnspython==2.4.2
whois==0.9.27
paramiko==3.3.1
scapy==2.5.0
beautifulsoup4==4.12.2
selenium==4.15.2
pysocks==1.7.1
stem==1.8.2
sqlcipher3==0.5.1
argon2-cffi==23.1.0
reportlab==4.0.7
jinja2==3.1.2
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Security-specific dependencies
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
bcrypt==4.1.2
itsdangerous==2.1.2
python-dateutil==2.8.2
websockets==12.0
asyncio-mqtt==0.13.0
aiodns==3.1.1
cchardet==2.1.7
python-magic==0.4.27
yara-python==4.5.0
ssdeep==3.4
pycrypto==2.6.1
pynacl==1.5.0
keyring==24.3.0
keyrings.alt==5.0.0
SecretStorage==3.3.3
dbus-python==1.3.2
jsonschema==4.20.0
marshmallow==3.20.2
cerberus==1.3.5
schema==0.7.5
voluptuous==0.14.1
click==8.1.7
rich==13.7.0
typer==0.9.0
httpx==0.25.2
starlette==0.27.0
slowapi==0.1.9
limits==3.7.0
flask-limiter==3.5.0
cachetools==5.3.2
memory-profiler==0.61.0
psutil==5.9.6
pymongo==4.6.0
elasticsearch==8.11.1
redis-py-cluster==2.1.3
prometheus-client==0.19.0
statsd==4.0.1
opencensus==0.11.4
opencensus-ext-prometheus==0.3.2
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation==0.43b0
structlog==23.2.0
loguru==0.7.2
colorlog==6.8.0
python-json-logger==2.0.7
fluent-logger==0.10.0
graypy==2.1.0
sentry-sdk==1.39.2
rollbar==0.16.3
bugsnag==4.6.1
newrelic==9.3.0
datadog==0.49.1
elastic-apm==6.20.0
jaeger-client==4.8.0
zipkin==0.2.1
py-spy==0.3.14
line-profiler==4.1.1
objgraph==3.6.0
pympler==0.9
tracemalloc==1.0
faulthandler==3.2
hunter==3.6.1
pdb==0.1
ipdb==0.13.13
pudb==2023.1
remote-pdb==2.1.0
werkzeug==3.0.1
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-restx==1.3.0
flask-marshmallow==0.15.0
flask-httpauth==4.8.0
flask-principal==0.4.0
flask-security==5.3.2
flask-login==0.6.3
flask-session==0.5.0
flask-caching==2.1.0
flask-compress==1.14
flask-talisman==1.1.0
flask-seasurf==1.1.2
flask-wtf==1.2.1
wtforms==3.1.1
email-validator==2.1.0
phonenumbers==8.13.26
python-magic==0.4.27
pillow==10.1.0
wand==0.6.13
opencv-python==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
pypdf2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
python-pptx==0.6.23
markdown==3.5.1
bleach==6.1.0
html5lib==1.1
lxml==4.9.3
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2
chromedriver-autoinstaller==0.6.2
geckodriver-autoinstaller==0.1.0
playwright==1.40.0
pyppeteer==1.0.2
requests-html==0.10.0
mechanicalsoup==1.3.0
urllib3==2.1.0
httplib2==0.22.0
tornado==6.4
twisted==23.10.0
gevent==23.9.1
eventlet==0.33.3
greenlet==3.0.1
gunicorn==21.2.0
waitress==2.1.2
mod-wsgi==5.0.0
cheroot==10.0.0
bjoern==3.2.2
meinheld==1.0.2
falcon==3.1.1
bottle==0.12.25
cherrypy==18.8.0
pyramid==2.0.2
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-silk==5.0.4
django-compressor==4.4
django-storages==1.14.2
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-environ==0.11.2
django-configurations==2.5
django-model-utils==4.3.1
django-mptt==0.15.0
django-taggit==5.0.1
django-guardian==2.4.0
django-rules==3.3.0
django-fsm==2.8.1
django-polymorphic==3.1.0
django-countries==7.5.1
django-phonenumber-field==7.2.0
django-timezone-field==6.0.1
django-jsonfield==3.1.0
django-picklefield==3.1.0
django-encrypted-model-fields==0.6.5
django-cryptography==1.1
django-fernet-fields==0.6
django-pgcrypto-fields==2.5.1
django-braces==1.15.0
django-extra-views==0.14.0
django-tables2==2.7.0
django-bootstrap4==23.2
django-widget-tweaks==1.5.0
django-crispy-bootstrap4==2023.1
django-activity-stream==2.0.0
django-notifications-hq==1.8.3
django-channels==4.0.0
channels-redis==4.1.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
djangorestframework-csv==3.0.2
djangorestframework-jsonapi==6.1.0
djangorestframework-gis==1.0
djangorestframework-bulk==0.2.1
djangorestframework-filters==1.0.0.dev2
djangorestframework-guardian==0.3.0
djangorestframework-recursive==0.1.2
drf-spectacular==0.27.0
drf-yasg==1.21.7
swagger-spec-validator==3.0.3
flex==6.14.1
apispec==6.3.0
apispec-webframeworks==0.5.2
flasgger==0.9.7.1
connexion==3.0.3
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
jsonref==1.1.0
prance==23.6.21.0
bravado==11.0.3
bravado-core==6.1.1
pyswagger==0.8.39
swagger-parser==1.0.0
swagger-tester==0.2.9
dredd-hooks==0.2.0
prism-cli==0.1.0
spectral==0.1.0
redoc-cli==0.2.0
swagger-codegen==3.0.0
openapi-generator==7.1.0
autorest==3.0.6369
quicktype==0.1.0
json-schema-codegen==0.1.0
datamodel-code-generator==0.25.1
pydantic-generators==0.3.0
sqlalchemy-codegen==0.1.0
alembic-autogenerate-enums==0.1.0
marshmallow-sqlalchemy==0.29.0
sqlalchemy-utils==0.41.1
sqlalchemy-migrate==0.13.0
sqlalchemy-json==0.7.0
sqlalchemy-searchable==1.5.4
sqlalchemy-continuum==1.4.0
sqlalchemy-history==0.3.0
sqlalchemy-temporal==0.1.0
sqlalchemy-audit==0.1.0
sqlalchemy-logging==0.1.0
sqlalchemy-filters==0.13.0
sqlalchemy-pagination==0.0.2
sqlalchemy-repr==0.1.0
sqlalchemy-mixins==2.0.3
sqlalchemy-fancy-core==0.1.0
sqlalchemy-dict==0.1.0
sqlalchemy-imageattach==1.1.0
sqlalchemy-media==0.20.0
sqlalchemy-file==0.5.0
sqlalchemy-mptt==0.2.5
sqlalchemy-tree==0.3.0
sqlalchemy-nested-sets==0.1.0
sqlalchemy-ordered-model==0.1.0
sqlalchemy-sortable==0.1.0
sqlalchemy-defaults==0.1.0
sqlalchemy-easy-profile==0.1.0
sqlalchemy-stubs==0.4
mypy-extensions==1.0.0
typing-extensions==4.8.0
types-requests==2.31.0.10
types-urllib3==1.26.25.14
types-PyYAML==6.0.12.12
types-python-dateutil==2.8.19.14
types-setuptools==69.0.0.0
types-six==1.16.21.9
types-beautifulsoup4==4.12.0.7
types-lxml==2023.10.21
types-Pillow==10.1.0.2
types-redis==4.6.0.11
types-psutil==5.9.5.17
types-paramiko==3.3.0.0
types-cryptography==3.3.23.2
types-pyOpenSSL==23.3.0.0
types-PyJWT==1.7.1
types-passlib==1.7.7.13
types-bcrypt==3.2.0.2
types-argon2-cffi==21.3.2.2
types-itsdangerous==1.1.6
types-Werkzeug==1.0.9
types-Flask==1.1.6
types-Flask-SQLAlchemy==2.5.9.7
types-click==7.1.8
types-rich==13.0.3
types-typer==0.9.0.3
types-httpx==0.25.2
types-starlette==0.27.0.6
types-uvicorn==0.24.0.0
types-fastapi==0.104.1
types-pydantic==2.5.0.1
types-sqlalchemy==1.4.53.38
types-alembic==1.12.1.2
types-celery==5.3.4.7
types-redis-py-cluster==2.1.3.20230315
types-prometheus-client==0.19.0.0
types-statsd==4.0.1.2
types-opencensus==0.11.4.20230315
types-opentelemetry-api==1.22.0.0
types-structlog==23.2.0.0
types-loguru==0.7.2.0
types-colorlog==6.8.0.0
types-sentry-sdk==1.39.2.0
types-rollbar==0.16.3.0
types-bugsnag==4.6.1.0
types-newrelic==9.3.0.0
types-datadog==0.49.1.0
types-elastic-apm==6.20.0.0
types-jaeger-client==4.8.0.0

# Knowledge Graph Dependencies
scipy==1.11.4
python-louvain==0.16
leidenalg==0.10.1
igraph==0.11.3
pillow==10.1.0
python-igraph==0.11.3

# RL Agent Additional Dependencies
plotly==5.17.0
tensorboard==2.15.1
wandb==0.16.0
optuna==3.4.0
ray[tune]==2.8.0
imageio==2.31.6
psutil==5.9.6
tqdm==4.66.1
cloudpickle==3.0.0
tensorboard-data-server==0.7.2
protobuf==4.25.1